Use getDollarValue for ProductCard, move it to common.js.
