+++ components/rewards/Transaction.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Colors from '../../assets/Colors';
-import { displayDollarValue } from '../../lib/rewardsUtils';
+import { displayDollarValue } from '../../lib/common';
 import { Card, ContentContainer } from '../../styled/transaction';
 import { Caption, Subhead } from '../BaseComponents';
 import CircleIcon from '../CircleIcon';
