+++ lib/common.js
@@ -18,3 +18,11 @@ const BASE = new Airtable({
 
 export default BASE;
 export const IMG_KEY = process.env.IMG_API_KEY;
+
+export function displayDollarValue(amount, positive = true) {
+  const dollarValue = '$'.concat(amount.toFixed(2).toString());
+  if (positive) {
+    return dollarValue;
+  }
+  return '-'.concat(dollarValue);
+}
