+++ lib/rewardsUtils.js
@@ -17,13 +17,6 @@ function createTransactionData(record) {
     totalSale: transaction['Total Price']
   };
 }
-export function displayDollarValue(amount, positive = true) {
-  const dollarValue = '$'.concat(amount.toFixed(2).toString());
-  if (positive) {
-    return dollarValue;
-  }
-  return '-'.concat(dollarValue);
-}
 
 // Calls Airtable API to return a promise that
 // will resolve to be a user record.
